feat: Restructure project for improved modularity and Docker efficiency
- Create a `.dockerignore` file to exclude unnecessary files from Docker build context, enhancing build performance.
- Modify `.vscode/launch.json` to launch the Flask application via the `flask run` command instead of directly running the script, improving development environment consistency.
- Update `Dockerfile` to copy the application into a subdirectory, reflecting the new project structure.
- Simplify `Pipfile` and `requirements.txt` by removing unused dependencies, reducing the project's footprint.
- Refactor the Flask application structure:
  - Delete `src/app.py` and migrate its functionality into a new package structure within `src/app`.
  - Introduce `src/app/__init__.py` for application factory pattern implementation.
  - Add configuration management in `src/app/config.py`.
  - Organize endpoints into modular routes under `src/app/routes`.
  - Implement utility functions for GCP interactions in `src/app/utils/gcp_utils.py`.
- Add a new entry point `src/run.py` for the Flask application, aligning with the Flask application factory pattern.
