+++ Pipfile
@@ -7,13 +7,9 @@ name = "pypi"
 flask = "*"
 flask-cors = "*"
 flask-restx = "*"
-flask-dance = "*"
-google-cloud = "*"
-google-cloud-secret-manager = "*"
-python-dotenv = "*"
-gunicorn = "*"
 google-cloud-storage = "*"
 google-cloud-discoveryengine = "*"
+python-dotenv = "*"
 
 [dev-packages]
 black = "*"
