+++ requirements.txt
@@ -8,22 +8,17 @@ charset-normalizer==3.3.2; python_full_version >= '3.7.0'
 click==8.1.7; python_version >= '3.7'
 flask==3.0.2; python_version >= '3.8'
 flask-cors==4.0.0
-flask-dance==7.1.0; python_version >= '3.6'
 flask-restx==1.3.0
 google-api-core[grpc]==2.18.0; python_version >= '3.7'
 google-auth==2.29.0; python_version >= '3.7'
-google-cloud==0.34.0
 google-cloud-core==2.4.1; python_version >= '3.7'
 google-cloud-discoveryengine==0.11.10; python_version >= '3.7'
-google-cloud-secret-manager==2.19.0; python_version >= '3.7'
 google-cloud-storage==2.16.0; python_version >= '3.7'
 google-crc32c==1.5.0; python_version >= '3.7'
 google-resumable-media==2.7.0; python_version >= '3.7'
 googleapis-common-protos[grpc]==1.63.0; python_version >= '3.7'
-grpc-google-iam-v1==0.13.0; python_version >= '3.7'
 grpcio==1.62.1
 grpcio-status==1.62.1
-gunicorn==21.2.0; python_version >= '3.5'
 idna==3.6; python_version >= '3.5'
 importlib-resources==6.4.0; python_version >= '3.8'
 itsdangerous==2.1.2; python_version >= '3.7'
@@ -31,8 +26,6 @@ jinja2==3.1.3; python_version >= '3.7'
 jsonschema==4.21.1; python_version >= '3.8'
 jsonschema-specifications==2023.12.1; python_version >= '3.8'
 markupsafe==2.1.5; python_version >= '3.7'
-oauthlib==3.2.2; python_version >= '3.6'
-packaging==24.0; python_version >= '3.7'
 proto-plus==1.23.0; python_version >= '3.6'
 protobuf==4.25.3; python_version >= '3.8'
 pyasn1==0.6.0; python_version >= '3.8'
@@ -41,9 +34,7 @@ python-dotenv==1.0.1; python_version >= '3.8'
 pytz==2024.1
 referencing==0.34.0; python_version >= '3.8'
 requests==2.31.0; python_version >= '3.7'
-requests-oauthlib==2.0.0; python_version >= '3.4'
 rpds-py==0.18.0; python_version >= '3.8'
 rsa==4.9; python_version >= '3.6' and python_version < '4'
 urllib3==2.2.1; python_version >= '3.8'
-urlobject==2.4.3
 werkzeug==3.0.1; python_version >= '3.8'
