+++ src/app/routes/greetings.py
@@ -1,0 +1,5 @@
+from flask_restx import Resource
+
+class Greetings(Resource):
+    def get(self):
+        return {"greeting": "Hello, world!"}
